requirements.txt
python-dotenv==0.9.1
spotipy==2.4.4
flask==1.0.2
flask_cors==3.0.4
